+++ src/Renderer/XmlRenderer.php
@@ -7,6 +7,7 @@
 
 namespace Zend\Expressive\Hal\Renderer;
 
+use DateTimeInterface;
 use DOMDocument;
 use DOMNode;
 use Zend\Expressive\Hal\HalResource;
@@ -112,6 +113,11 @@ class XmlRenderer implements RendererInterface
             return $doc->createElement($name, $data);
         }
 
+        if (is_object($data)) {
+            $data = $this->createDataFromObject($data);
+            return $doc->createElement($name, $data);
+        }
+
         if (! is_array($data)) {
             throw Exception\InvalidResourceValueException::fromValue($data);
         }
@@ -141,5 +147,26 @@ class XmlRenderer implements RendererInterface
         }
 
         return $node;
+    }
+
+    /**
+     * @todo Detect JsonSerializable, and pass to
+     *     json_decode(json_encode($object), true), passing the final value
+     *     back to createResourceElement()?
+     * @param object $object
+     * @throws Exception\InvalidResourceValueException if unable to serialize
+     *     the data to a string.
+     */
+    private function createDataFromObject($object) : string
+    {
+        if ($object instanceof DateTimeInterface) {
+            return $object->format('c');
+        }
+
+        if (! method_exists($object, '__toString')) {
+            throw Exception\InvalidResourceValueException::fromObject($object);
+        }
+
+        return (string) $object;
     }
 }